fix(engine): long-audit follow-ups — LP64 guard by data model + rendcrow element type
Review follow-ups to the F1–F6 long-width audit (9375f3f):

- pigtool.h / pigtypes.h: key the int32 width guard on __LP64__ instead of
  __LINUX__/__ANDROID__. iOS/macOS arm64 is LP64 but defines WF_TARGET_IOS
  (not __LINUX__/__ANDROID__), so it fell through to `signed long` and got a
  64-bit int32 — the exact bug F1/F2 fix for Linux, latent on the 2026-ios
  port. Data model, not OS name, is what makes `long` 64-bit; ILP32 (PSX/
  x86-32) and LLP64 (Win64) stay on the `long`==32 branch unchanged.
- rendcrow.cc: _nTextures divisor sizeof(int32) → sizeof(packedAssetID), the
  on-disk element type. 4 bytes either way, but it names what the bytes are
  (sizeof(_texture[0]) would be wrong — Texture* is 8 bytes on LP64).
- oad.h FIXED32 / particle.cc range comments: (long) → (int32) tidy-up; no
  build impact (FIXED32 is unused in compiled C++; particle hunks are inside
  comments only).
- docs plan + TODO: record the iOS/rendcrow follow-ups and log two deferred
  cleanups — FourCC as a first-class type, and making the fixed-point math's
  64-bit intermediates explicitly int64.

+++ wfsource/source/oas/oad.h
@@ -7,7 +7,7 @@
 #define OAD_H
 
 #define FIXED16(n) ( (short)(n * 256) )
-#define FIXED32(n) ( (long)(n * 65536) )
+#define FIXED32(n) ( (int32)(n * 65536) )
 #define RADIOBUTTONNAMELEN 10
